back-end/requirements.txt
absl-py==1.4.0
aiohttp==3.8.4
aiosignal==1.3.1
astunparse==1.6.3
async-timeout==4.0.2
attrs==23.1.0
beautifulsoup4==4.12.2
cachetools==5.3.1
certifi==2023.5.7
charset-normalizer==3.1.0
dataclasses-json==0.5.7
flatbuffers==23.5.26
frozenlist==1.3.3
fsspec==2023.6.0
gast==0.4.0
google-auth==2.19.1
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
greenlet==2.0.2
grpcio==1.54.2
h5py==3.8.0
idna==3.4
jaraco.context==4.3.0
jax==0.4.12
keras==2.12.0
langchain==0.0.195
langchainplus-sdk==0.0.8
libclang==16.0.0
llama-index==0.6.21.post1
Markdown==3.4.3
MarkupSafe==2.1.3
marshmallow==3.19.0
marshmallow-enum==1.5.1
ml-dtypes==0.2.0
more-itertools==9.1.0
multidict==6.0.4
mypy-extensions==1.0.0
numexpr==2.8.4
numpy==1.23.5
oauthlib==3.2.2
openai==0.27.8
openapi-schema-pydantic==1.2.4
opt-einsum==3.3.0
packaging==23.1
pandas==2.0.2
Pillow==9.5.0
protobuf==4.23.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydantic==1.10.9
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
regex==2023.6.3
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
scipy==1.10.1
six==1.16.0
soupsieve==2.4.1
SQLAlchemy==2.0.15
tenacity==8.2.2
tensorflow==2.12.0
termcolor==2.3.0
tiktoken==0.4.0
tqdm==4.65.0
typing-inspect==0.8.0
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.16
Werkzeug==2.3.6
wikipedia==1.4.0
wolframalpha==5.0.0
wrapt==1.14.1
xmltodict==0.13.0
yarl==1.9.2